Optimize to make page transition faster #28
Resolve #28
