+++ apps/dailyscrum/next.config.mjs
@@ -23,6 +23,11 @@ const nextConfig = {
       },
     ];
   },
+  logging: {
+    fetches: {
+      fullUrl: true,
+    },
+  },
 };
 
 export default nextConfig;
